fix(temper-agents): drop stale machete ignore + token_budget u64→u32
Consolidated-review findings. F1: remove [package.metadata.cargo-machete]
ignore block — temper-next and ts-rs are both genuinely used now, so the
ignore suppressed nothing. F2: token_budget u64→u32 so the generated TS is
`number | null` (matching the JSON wire value) instead of `bigint | null`;
u32 (~4.29B) is ample for a token budget. Spec code block amended; TS
regenerated.

+++ crates/temper-agents/src/profile.rs
@@ -50,7 +50,7 @@ pub struct DeploymentProfile {
     pub residency: Residency,
     /// Token-denominated budget — neither runtime exposes a managed dollar
     /// budget, so spend governance is expressed in tokens. `None` = runtime default.
-    pub token_budget: Option<u64>,
+    pub token_budget: Option<u32>,
 }
 
 #[cfg(test)]
